style: split file user controller
